processing/requirements.txt
# Processing requirements
pyspark==3.5.1
boto3==1.34.69
botocore==1.34.69
great-expectations==0.18.12
psycopg2-binary==2.9.9
sqlalchemy==1.4.52
pandas==2.2.1
pyarrow==15.0.2
pytest==8.1.1
pytest-cov==5.0.0
chispa==0.9.4
flake8==7.0.0
mypy==1.9.0
